move getMailBodyText() and isDraft() function to CommonMailUtils.ts
